fix(deploy): add next.config.js, set up standard layout for app, and reorganize file structure
