+++ next.config.js
@@ -1,0 +1,11 @@
+// next.config.js
+
+/** @type { import('next').NextConfig} */
+
+const nextConfig = {
+  reactStrictMode: true,
+
+  /* config options here */
+};
+
+module.exports = nextConfig;
